+++ src/entities/post/index.server.ts
@@ -1,2 +1,2 @@
-export { getPostServer, getProfilePostsServer } from './api/postsApi.server'
+export { getPostServer, getProfilePostServer, getProfilePostsServer } from './api/postsApi.server'
 export type { Post, PostsPage } from './model/types'